+++ src/main/java/ps/repository/EstudianteRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import ps.model.Estudiante;
 
 @Repository
-public interface EstudianteRepository extends JpaRepository<Estudiante, Long>{
+public interface EstudianteRepository extends JpaRepository<Estudiante, Integer>{
 
 }
